app/pyproject.toml
[tool.poetry]
name = "fixmaster"
version = "0.1.0"
description = ""
authors = ["nariman079 <nariman079i@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
celery = "^5.4.0"
redis = "^5.0.4"
python-dotenv = "^1.0.1"
telebot = "^0.0.5"
drf-spectacular = "^0.27.2"
postgres = "^4.0"
django-celery-beat = "^2.6.0"
drf-yasg = "^1.21.7"
pillow = "^10.3.0"
uvicorn = "^0.29.0"
drf-spectacular-sidecar = "^2024.5.1"
ruff = "^0.12.1"
django-prometheus = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
